Fix typo ('two'=>'three') in Traversable laws
The comment that precedes the Traversable type
class definition declares there are two laws,
but then goes on to define three (naturality,
identity, composition).

Typeclassopedia does suggest that identity and
composition are the traversable laws and that
naturality follows if you are dealing with
applicatives. But I definitely don't know enough
to assess the truth of that claim.
(ref: https://wiki.haskell.org/Typeclassopedia#Laws_7)

Either way the current comment seems misleading.